chore: refactor file names
- change file names from kebab case to camel case to follow the expected convention
