+++ routes/users.js
@@ -1,6 +1,6 @@
 import express from 'express';
 import UsersController from '../controllers/users';
-import { handleRegistration } from '../middlewares/handle-validation';
+import { handleRegistration } from '../middlewares/handleValidation';
 import Validator from '../middlewares/validator';
 
 const router = express.Router();
